requirements.txt
djangorestframework
git+git://github.com/gkuhn1/django-admin-multiupload.git
django-filter
gunicorn

